requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
datasets>=2.14.0
huggingface-hub>=0.16.0
opencv-python>=4.8.0
mediapipe>=0.10.0
scikit-learn>=1.3.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
pyyaml>=6.0
click>=8.1.0
pandas>=2.0.0
# Later additions:
# lerobot (install from git when ready)
# depth-anything-v2 (if we add depth)
# rerun-sdk (for visualization)
